Persist task queue state and fix offline pytest

+++ api/v1/schemas/system_config.py
@@ -30,7 +30,7 @@ class SystemConfigFieldSchema(BaseModel):
     default_value: Optional[str] = None
     options: List[str | SystemConfigOption] = Field(default_factory=list)
     validation: Dict[str, Any] = Field(default_factory=dict)
-    display_order: int
+    display_order: float
 
 
 class SystemConfigCategorySchema(BaseModel):
